refactor(server): :recycle: Add index file for twilio connector
